+++ decidim-accountability/app/forms/decidim/accountability/admin/result_form.rb
@@ -25,6 +25,8 @@ module Decidim
 
         validates :title, translatable_presence: true
 
+        validates :progress, numericality: { greater_than_or_equal_to: 0, less_than_or_equal_to: 100 }, if: ->(form) { form.progress.present? }
+
         validates :scope, presence: true, if: ->(form) { form.decidim_scope_id.present? }
         validates :category, presence: true, if: ->(form) { form.decidim_category_id.present? }