+++ decidim-core/lib/decidim/content_block_manifest.rb
@@ -12,7 +12,7 @@ module Decidim
   #
   # Content blocks are intended to be used in the home page, for example.
   #
-  # A content block has a set of options and an associated `cell` that will
+  # A content block has a set of settings and an associated `cell` that will
   # handle the layout logic. They can also have attached images that can be used
   # as background images, for example. You must explicitly specify the number of
   # images the block will have (this means the number of attached images cannot
@@ -23,13 +23,12 @@ module Decidim
     include Virtus.model
 
     attribute :name, Symbol
+    attribute :i18n_name_key, String
     attribute :cell_name, String, writer: :private
     attribute :image_names, Array[Symbol]
-    attribute :options, Array[Hash]
 
-    validates :name, :cell_name, presence: true
+    validates :name, :cell_name, :i18n_name_key, presence: true
     validate :image_names_are_unique
-    validate :option_names_are_unique
 
     # Public: Registers an image with a given name. Use `#images` to retrieve
     # them all.
@@ -45,27 +44,28 @@ module Decidim
       self.cell_name = cell_name
     end
 
-    # Public: Registers an option. Use `#options` to retrieve them all.
-    def option(name, type, metadata = {})
-      raise OptionNameCannotBeBlank, "Option names cannot be blank" if name.blank?
-      raise OptionTypeCannotBeBlank, "Option types cannot be blank" if type.blank?
+    # Public: Registers the I18n key this contnt block will use to retrieve its
+    # public name. Use `#i18n_name_key` to retrieve it.
+    def public_name_key(i18n_key)
+      self.i18n_name_key = i18n_key
+    end
 
-      options << { name: name, type: type }.merge(metadata)
+    def has_settings?
+      settings.attributes.any?
+    end
+
+    def settings(&block)
+      @settings ||= SettingsManifest.new
+      yield(@settings) if block
+      @settings
     end
 
     private
-
-    def option_names_are_unique
-      option_names = options.map { |option| option.fetch(:name) }
-      errors.add(:options, :invalid) if option_names.count != option_names.uniq.count
-    end
 
     def image_names_are_unique
       errors.add(:image_names, :invalid) if image_names.count != image_names.uniq.count
     end
 
     class ImageNameCannotBeBlank < StandardError; end
-    class OptionNameCannotBeBlank < StandardError; end
-    class OptionTypeCannotBeBlank < StandardError; end
   end
 end