+++ decidim-proposals/spec/system/proposals_spec.rb
@@ -122,20 +122,20 @@ describe "Proposals", type: :system do
 
     context "when a proposal has been linked in a result" do
       let(:proposal) { create(:proposal, component: component) }
-      let(:dummy_component) do
-        create(:component, manifest_name: :dummy, participatory_space: proposal.component.participatory_space)
+      let(:accountability_component) do
+        create(:component, manifest_name: :accountability, participatory_space: proposal.component.participatory_space)
       end
-      let(:dummy_resource) { create(:dummy_resource, component: dummy_component) }
+      let(:result) { create(:result, component: accountability_component) }
 
       before do
-        dummy_resource.link_resources([proposal], "included_proposals")
+        result.link_resources([proposal], "included_proposals")
       end
 
       it "shows related resources" do
         visit_component
         click_link proposal.title
 
-        expect(page).to have_i18n_content(dummy_resource.title)
+        expect(page).to have_i18n_content(result.title)
       end
     end